feat: home-manager 'profiles.<misc>' -> 'jk.<misc>'
so far this pattern is, a lot nicer than the whole 'profiles' business.

it's just a bunch of switches to turn on common groups of config
options, basically using 'options' as a way to collect things that had
previously been relegated to a bunch of imports.

next up: repeat this pattern for the system-level config.